+++ src/business_logic/transaction/objects/internal_deploy_account.rs
@@ -1,0 +1,254 @@
+use felt::Felt;
+use num_traits::Zero;
+
+use crate::{
+    business_logic::{
+        execution::{
+            error::ExecutionError,
+            execution_entry_point::ExecutionEntryPoint,
+            objects::{CallInfo, TransactionExecutionContext, TransactionExecutionInfo},
+        },
+        fact_state::{contract_state::StateSelector, state::ExecutionResourcesManager},
+        state::state_api::{State, StateReader},
+    },
+    core::{
+        errors::{state_errors::StateError, syscall_handler_errors::SyscallHandlerError},
+        transaction_hash::starknet_transaction_hash::calculate_deploy_account_transaction_hash,
+    },
+    definitions::{
+        constants::{CONSTRUCTOR_ENTRY_POINT_SELECTOR, VALIDATE_DEPLOY_ENTRY_POINT_SELECTOR},
+        general_config::{StarknetChainId, StarknetGeneralConfig},
+        transaction_type::TransactionType,
+    },
+    hash_utils::calculate_contract_address,
+    services::api::contract_class::{ContractClass, EntryPointType},
+    utils::{calculate_tx_resources, Address},
+};
+
+use super::internal_invoke_function::verify_no_calls_to_other_contracts;
+
+#[derive(Clone, Debug)]
+pub struct InternalDeployAccount {
+    contract_address: Address,
+    contract_address_salt: Address,
+    class_hash: [u8; 32],
+    constructor_calldata: Vec<Felt>,
+    version: u64,
+    nonce: u64,
+    max_fee: u64,
+    signature: Vec<Felt>,
+    chain_id: StarknetChainId,
+}
+
+impl InternalDeployAccount {
+    #![allow(unused)] // TODO: delete once used
+    #[allow(clippy::too_many_arguments)]
+    pub fn new(
+        class_hash: [u8; 32],
+        max_fee: u64,
+        version: u64,
+        nonce: u64,
+        constructor_calldata: Vec<Felt>,
+        signature: Vec<Felt>,
+        contract_address_salt: Address,
+        chain_id: StarknetChainId,
+    ) -> Result<Self, SyscallHandlerError> {
+        let contract_address = calculate_contract_address(
+            &contract_address_salt,
+            &Felt::from_bytes_be(&class_hash),
+            &constructor_calldata,
+            Address(Felt::zero()),
+        )?;
+
+        Ok(Self {
+            contract_address: Address(contract_address),
+            contract_address_salt,
+            class_hash,
+            constructor_calldata,
+            version,
+            nonce,
+            max_fee,
+            signature,
+            chain_id,
+        })
+    }
+
+    pub fn get_state_selector(&self, _general_config: StarknetGeneralConfig) -> StateSelector {
+        StateSelector {
+            contract_addresses: vec![self.contract_address.clone()],
+            class_hashes: vec![self.class_hash],
+        }
+    }
+
+    fn _apply_specific_concurrent_changes<S>(
+        &self,
+        state: &mut S,
+        general_config: &StarknetGeneralConfig,
+    ) -> Result<TransactionExecutionInfo, StateError>
+    where
+        S: Default + State + StateReader,
+    {
+        let contract_class = state.get_contract_class(&self.class_hash)?;
+
+        state.deploy_contract(self.contract_address.clone(), self.class_hash)?;
+
+        let mut resources_manager = ExecutionResourcesManager::default();
+        let constructor_call_info = self
+            .handle_constructor(
+                contract_class,
+                state,
+                general_config,
+                &mut resources_manager,
+            )
+            .map_err::<StateError, _>(|_| todo!())?;
+
+        let validate_info = self
+            .run_validate_entrypoint(state, &mut resources_manager, general_config)
+            .map_err::<StateError, _>(|_| todo!())?;
+
+        let actual_resources = calculate_tx_resources(
+            resources_manager,
+            &[Some(constructor_call_info.clone()), validate_info.clone()],
+            TransactionType::DeployAccount,
+            state.count_actual_storage_changes(),
+            None,
+        )
+        .map_err::<StateError, _>(|_| todo!())?;
+
+        Ok(
+            TransactionExecutionInfo::create_concurrent_stage_execution_info(
+                validate_info,
+                Some(constructor_call_info),
+                actual_resources,
+                Some(TransactionType::DeployAccount),
+            ),
+        )
+    }
+
+    pub fn handle_constructor<S>(
+        &self,
+        contract_class: ContractClass,
+        state: &mut S,
+        general_config: &StarknetGeneralConfig,
+        resources_manager: &mut ExecutionResourcesManager,
+    ) -> Result<CallInfo, ExecutionError>
+    where
+        S: Default + State + StateReader,
+    {
+        let num_constructors = contract_class
+            .entry_points_by_type
+            .get(&EntryPointType::Constructor)
+            .map(Vec::len)
+            .unwrap_or(0);
+
+        match num_constructors {
+            0 => {
+                if !self.constructor_calldata.is_empty() {
+                    todo!()
+                }
+
+                Ok(CallInfo::empty_constructor_call(
+                    self.contract_address.clone(),
+                    Address(Felt::zero()),
+                    Some(self.class_hash),
+                ))
+            }
+            _ => self.run_constructor_entrypoint(state, general_config, resources_manager),
+        }
+    }
+
+    pub fn run_constructor_entrypoint<S>(
+        &self,
+        state: &mut S,
+        general_config: &StarknetGeneralConfig,
+        resources_manager: &mut ExecutionResourcesManager,
+    ) -> Result<CallInfo, ExecutionError>
+    where
+        S: Default + State + StateReader,
+    {
+        let call = ExecutionEntryPoint::new(
+            self.contract_address.clone(),
+            self.constructor_calldata.clone(),
+            CONSTRUCTOR_ENTRY_POINT_SELECTOR.clone(),
+            Address(Felt::zero()),
+            EntryPointType::Constructor,
+            None,
+            None,
+        );
+
+        let call_info = call.execute(
+            state,
+            general_config,
+            resources_manager,
+            &self.get_execution_context(general_config.validate_max_n_steps),
+        )?;
+
+        verify_no_calls_to_other_contracts(&call_info)
+            .map_err(|_| ExecutionError::OtherContractCalls)?;
+        Ok(call_info)
+    }
+
+    pub fn get_execution_context(&self, n_steps: u64) -> TransactionExecutionContext {
+        TransactionExecutionContext::new(
+            self.contract_address.clone(),
+            calculate_deploy_account_transaction_hash(
+                self.version,
+                self.contract_address.clone(),
+                Felt::from_bytes_be(&self.class_hash),
+                &self.constructor_calldata,
+                self.max_fee,
+                self.nonce,
+                self.contract_address_salt.0.clone(),
+                self.chain_id.to_felt(),
+            )
+            .unwrap(),
+            self.signature.clone(),
+            self.max_fee,
+            self.nonce.into(),
+            n_steps,
+            self.version,
+        )
+    }
+
+    pub fn run_validate_entrypoint<S>(
+        &self,
+        state: &mut S,
+        resources_manager: &mut ExecutionResourcesManager,
+        general_config: &StarknetGeneralConfig,
+    ) -> Result<Option<CallInfo>, ExecutionError>
+    where
+        S: Default + State + StateReader,
+    {
+        if self.version == 0 {
+            return Ok(None);
+        }
+
+        let call = ExecutionEntryPoint::new(
+            self.contract_address.clone(),
+            [
+                Felt::from_bytes_be(&self.class_hash),
+                self.contract_address_salt.0.clone(),
+            ]
+            .into_iter()
+            .chain(self.constructor_calldata.iter().cloned())
+            .collect(),
+            VALIDATE_DEPLOY_ENTRY_POINT_SELECTOR.clone(),
+            Address(Felt::zero()),
+            EntryPointType::External,
+            None,
+            None,
+        );
+
+        let call_info = call.execute(
+            state,
+            general_config,
+            resources_manager,
+            &self.get_execution_context(general_config.validate_max_n_steps),
+        )?;
+
+        verify_no_calls_to_other_contracts(&call_info)
+            .map_err(|_| ExecutionError::OtherContractCalls)?;
+
+        Ok(Some(call_info))
+    }
+}
