+++ src/business_logic/transaction/objects/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod interal_declare;
 pub(crate) mod internal_deploy;
+pub(crate) mod internal_deploy_account;
 pub(crate) mod internal_invoke_function;
